refactor(viewer): extract shared color legend into color_legend.js
Consolidate the duplicated gradient bar + label logic from heatmap.js
and histogram_heatmap.js into a single color_legend.js module:

- buildGradientCanvas(colorFn): renders a gradient bar from a 0..1
  color function (used by both inferno and viridis palettes)
- colorArrayToFn(colors): converts a color array to a color function
- ensureLegendBar(wrapper, canvas, extras): creates or reuses a flex
  container with [minLabel] [bar] [maxLabel] + optional extra elements
- Shared constants: BAR_WIDTH, BAR_HEIGHT, BAR_TOP, LABEL_GAP

Both heatmap types now import from this shared module. The histogram
heatmap passes its checkbox as an extra element.
